Moves Controls to components folder
