+++ src/components/Controls/Controls.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
-import AddPalette from '../AddPalette/AddPalette'
-import CreateProject from '../CreateProject/CreateProject'
+import AddPalette from '../../containers/AddPalette/AddPalette'
+import CreateProject from '../../containers/CreateProject/CreateProject'
 import PropTypes from 'prop-types'
 
 class Controls extends Component {
