+++ src/main.scss
@@ -2,7 +2,7 @@
 @import './components/Swatch/Swatch.scss';
 @import './containers/App/App.scss';
 @import './containers/AddPalette/AddPalette.scss';
-@import './containers/Controls/Controls.scss';
+@import './components/Controls/Controls.scss';
 @import './containers/CreateProject//CreateProject.scss';
 @import './containers/Generator/Generator.scss';
 @import './components/Loading/Loading.scss';
